+++ src/ru/yandex/practicum/delivery/StandardParcel.java
@@ -3,7 +3,7 @@ package ru.yandex.practicum.delivery;
 public class StandardParcel extends Parcel {
     private static final int BASE_COST = 2;
 
-    public StandardParcel(String description, int weight, String deliveryAddress, byte sendDay) {
+    public StandardParcel(String description, int weight, String deliveryAddress, int sendDay) {
         super(description, weight, deliveryAddress, sendDay);
     }
