+++ test/ru/yandex/practicum/PerishableParcelTest.java
@@ -9,7 +9,7 @@ public class PerishableParcelTest {
     @Test
     void testNotExpired_StandardScenario() {
         PerishableParcel p = new PerishableParcel("Сыр", 2, "Воронеж",
-                (byte) 1, 5);
+                1, 5);
         assertFalse(p.isExpired(3),
                 "Посылка не должна быть просрочена, если текущий день 3, а срок до 6");
     }
@@ -17,7 +17,7 @@ public class PerishableParcelTest {
     @Test
     void testExpired_StandardScenario() {
         PerishableParcel p = new PerishableParcel("Йогурт", 1,
-                "Тула", (byte) 1, 2);
+                "Тула", 1, 2);
         assertTrue(p.isExpired(5),
                 "Посылка должна быть просрочена, если текущий день 5, а срок истек на 3-й день");
     }
@@ -25,7 +25,7 @@ public class PerishableParcelTest {
     @Test
     void testBoundaryCase_OnExpirationDay() {
         PerishableParcel p = new PerishableParcel("Молоко", 1,
-                "Казань", (byte) 1, 5);
+                "Казань", 1, 5);
         assertFalse(p.isExpired(6),
                 "Посылка не должна считаться просроченной ровно в день истечения срока");
     }
